Fix as suggested.
